feat: graphQl disable option
Task: PT-15055
