+++ src/main/java/com/infomaximum/platform/component/frontend/engine/service/graphqlrequestexecute/utils/GraphQLExecutionResultUtils.java
@@ -1,14 +1,26 @@
 package com.infomaximum.platform.component.frontend.engine.service.graphqlrequestexecute.utils;
 
 import com.infomaximum.cluster.graphql.executor.struct.GExecutionResult;
-import com.infomaximum.cluster.graphql.struct.GRequest;
+import com.infomaximum.cluster.graphql.executor.struct.GSourceLocation;
+import com.infomaximum.cluster.graphql.executor.struct.GSubscriptionPublisher;
+import com.infomaximum.platform.component.frontend.engine.service.graphqlrequestexecute.struct.GExecutionStatistics;
+import com.infomaximum.platform.component.frontend.engine.service.graphqlrequestexecute.struct.GraphQLResponse;
+import com.infomaximum.platform.exception.GraphQLWrapperPlatformException;
 import com.infomaximum.platform.exception.PlatformException;
 import com.infomaximum.platform.exception.runtime.PlatformRuntimeException;
 import com.infomaximum.platform.sdk.exception.GeneralExceptionBuilder;
+import com.infomaximum.platform.sdk.graphql.out.GOutputFile;
+import com.infomaximum.platform.utils.ExceptionUtils;
+import graphql.AssertException;
 import graphql.ExceptionWhileDataFetching;
 import graphql.GraphQLError;
+import graphql.execution.NonNullableValueCoercedAsNullException;
+import graphql.language.SourceLocation;
+import graphql.schema.CoercingParseValueException;
+import net.minidev.json.JSONObject;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -49,5 +61,67 @@ public class GraphQLExecutionResultUtils {
         }
 
         return errors.entrySet().stream().map(entry -> entry.getKey() + " (" + entry.getValue() + ")").collect(Collectors.joining(", "));
+    }
+
+    // Формирование ответа из результата
+    public static GraphQLResponse<JSONObject> buildResponse(GExecutionResult executionResult, GExecutionStatistics statistics) {
+        GOutputFile gOutputFile = findOutputFile(executionResult.getData());
+        if (gOutputFile != null) {
+            return new GraphQLResponse(gOutputFile, false, statistics);
+        } else {
+            if (executionResult.getData() instanceof GSubscriptionPublisher) {
+                return new GraphQLResponse(executionResult.getData(), false, statistics);
+            } else {
+                return new GraphQLResponse(new JSONObject(executionResult.getData()), false, statistics);
+            }
+        }
+    }
+
+    // Оборачиваем ошибку в обертку
+    public static GraphQLWrapperPlatformException coercionGraphQLPlatformException(Throwable throwable) {
+        if (throwable instanceof GraphQLWrapperPlatformException) {
+            return (GraphQLWrapperPlatformException) throwable;
+        } else if (throwable instanceof PlatformException) {
+            return new GraphQLWrapperPlatformException((PlatformException) throwable);
+        } else if (throwable instanceof PlatformRuntimeException) {
+            return new GraphQLWrapperPlatformException(((PlatformRuntimeException) throwable).getPlatformException());
+        } else if (throwable instanceof AssertException
+                || throwable instanceof CoercingParseValueException
+                || throwable instanceof NonNullableValueCoercedAsNullException
+        ) {
+            List<GSourceLocation> locations = null;
+            if (throwable instanceof GraphQLError graphQLError) {
+                locations = convertLocation(graphQLError);
+            }
+            return new GraphQLWrapperPlatformException(
+                    GeneralExceptionBuilder.buildGraphQLValidationException(throwable.getMessage()),
+                    locations,
+                    null
+            );
+        } else {
+            throw ExceptionUtils.coercionRuntimeException(throwable);
+        }
+    }
+
+    public static List<GSourceLocation> convertLocation(GraphQLError graphQLError) {
+        List<SourceLocation> nLocations = graphQLError.getLocations();
+        if (nLocations != null) {
+            return nLocations.stream().map(GSourceLocation::new).toList();
+        } else {
+            return null;
+        }
+    }
+
+    private static GOutputFile findOutputFile(Object response) {
+        if (response instanceof Map) {
+            for (Object oEntry : ((Map) response).entrySet()) {
+                Map.Entry<String, Object> entry = (Map.Entry<String, Object>) oEntry;
+                GOutputFile iFindOutputFile = findOutputFile(entry.getValue());
+                if (iFindOutputFile != null) return iFindOutputFile;
+            }
+        } else if (response instanceof GOutputFile) {
+            return (GOutputFile) response;
+        }
+        return null;
     }
 }
